pyproject.toml
[project]
name = "event-face-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "markdown>=3.9",
    "psycopg2>=2.9.10",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "django-stubs[mypy]>=5.2.5",
    "djangorestframework-stubs>=3.16.3",
    "mypy>=1.18.2",
    "mypy-extensions>=1.1.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.1",
]

[tool.ruff]
exclude = ["**/migrations/**"]

[tool.mypy]
exclude = ".*migrations.*"

plugins = ["mypy_django_plugin.main"]

[mypy.plugins.django-stubs]
django_settings_module = "src.core.settings"

